Docker branch is not needed (#1)
* NPM caching

By adding package.json in its own add step, we get to rebuild images from a cache. Since package.json hasn't
changed, docker won't rerun the command, or following ones, until it finds a step which invalidates the cache, most
likely a change inside app/.

* Bringing an actor into the mix

Whether this is correct or not, it means the reducer is solely interested in reducing the state to a known place, and
the component doesn't need to do anything with an onload, or similar.